+++ badger-optimiser/src/main.rs
@@ -11,10 +11,10 @@ use std::path::PathBuf;
 use std::process::exit;
 
 use clap::Parser;
-use tket2::json::{load_tk1_json_file, save_tk1_json_file};
 use tket2::optimiser::badger::log::BadgerLogger;
 use tket2::optimiser::badger::BadgerOptions;
 use tket2::optimiser::{BadgerOptimiser, DefaultBadgerOptimiser};
+use tket2::serialize::{load_tk1_json_file, save_tk1_json_file};
 
 #[cfg(feature = "peak_alloc")]
 #[global_allocator]
