+++ tket2-py/src/circuit.rs
@@ -16,8 +16,8 @@ use std::fmt;
 
 use hugr::{type_row, Hugr, HugrView, PortIndex};
 use tket2::extension::{LINEAR_BIT, REGISTRY};
-use tket2::json::TKETDecode;
 use tket2::rewrite::CircuitRewrite;
+use tket2::serialize::TKETDecode;
 use tket_json_rs::circuit_json::SerialCircuit;
 
 use crate::utils::create_py_exception;
@@ -83,7 +83,7 @@ create_py_exception!(
 );
 
 create_py_exception!(
-    tket2::json::OpConvertError,
+    tket2::serialize::pytket::OpConvertError,
     PyOpConvertError,
     "Error type for the conversion between tket2 and tket1 operations."
 );
