+++ tket2/src/serialize.rs
@@ -1,0 +1,9 @@
+//! Utilities for serializing circuits.
+//!
+//! See [`crate::serialize::pytket`] for serialization to and from the legacy pytket format.
+pub mod pytket;
+
+pub use pytket::{
+    load_tk1_json_file, load_tk1_json_reader, load_tk1_json_str, save_tk1_json_file,
+    save_tk1_json_str, save_tk1_json_writer, TKETDecode,
+};
