+++ tket2/src/serialize/pytket/decoder.rs
@@ -18,8 +18,8 @@ use tket_json_rs::circuit_json::SerialCircuit;
 
 use super::op::JsonOp;
 use super::{try_param_to_constant, METADATA_IMPLICIT_PERM, METADATA_PHASE};
+use super::{METADATA_B_REGISTERS, METADATA_Q_REGISTERS};
 use crate::extension::{LINEAR_BIT, REGISTRY};
-use crate::json::{METADATA_B_REGISTERS, METADATA_Q_REGISTERS};
 use crate::symbolic_constant_op;
 
 /// The state of an in-progress [`DFGBuilder`] being built from a [`SerialCircuit`].
